backend/requirements.txt
Django
django-filter==22.1
djangorestframework==3.14.0
djoser==2.1.0
drf-extra-fields==3.4.1
gunicorn
Pillow==9.2.0
psycopg2-binary
python-dotenv==0.21.0
reportlab==3.6.11
